+++ Code/CryGame/GameCVars.cpp
@@ -604,6 +604,7 @@ void SCVars::InitCVars(IConsole* pConsole)
 	pConsole->Register("mp_animationWeaponMultSpeed", &mp_animationWeaponMultSpeed, 3.0f, VF_NOT_NET_SYNCED);
 	pConsole->Register("mp_animationModelMult", &mp_animationModelMult, 1.0f, VF_NOT_NET_SYNCED);
 	pConsole->Register("mp_animationModelMultSpeed", &mp_animationModelMultSpeed, 1.3f, VF_NOT_NET_SYNCED);
+	pConsole->Register("mp_menuSpeed", &mp_menuSpeed, 1.0f, VF_NOT_NET_SYNCED);
 }
 
 //------------------------------------------------------------------------